move to dir and small typos
