+++ src/AntiCorruptionLayer/MessageFilters/AllowAllMessagesTest.php
@@ -2,8 +2,11 @@
 
 declare(strict_types=1);
 
-namespace EventSauce\EventSourcing\AntiCorruptionLayer;
+namespace EventSauce\EventSourcing\AntiCorruptionLayer\MessageFilters;
 
+use EventSauce\EventSourcing\AntiCorruptionLayer\StubExcludedEvent;
+use EventSauce\EventSourcing\AntiCorruptionLayer\StubPrivateEvent;
+use EventSauce\EventSourcing\AntiCorruptionLayer\StubPublicEvent;
 use EventSauce\EventSourcing\Message;
 use PHPUnit\Framework\TestCase;
