+++ src/AntiCorruptionLayer/Translators/MessageTranslatorPerPayloadType.php
@@ -2,10 +2,9 @@
 
 declare(strict_types=1);
 
-namespace EventSauce\EventSourcing\AntiCorruptionLayer;
+namespace EventSauce\EventSourcing\AntiCorruptionLayer\Translators;
 
 use EventSauce\EventSourcing\Message;
-
 use function get_class;
 
 class MessageTranslatorPerPayloadType implements MessageTranslator
